refactor: move the position of encoder and decoder
